+++ glide/src/main/java/com/bumptech/glide/load/engine/cache/MemoryCache.java
@@ -7,14 +7,10 @@ import com.bumptech.glide.load.engine.Resource;
 
 /**
  * An interface for adding and removing resources from an in memory cache.
- * <p>
- * 从内存缓存中添加和移除资源的接口
  */
 public interface MemoryCache {
     /**
      * An interface that will be called whenever a bitmap is removed from the cache.
-     * <p>
-     * 每当一个Bitmap从缓存中移除就会调用该接口
      */
     interface ResourceRemovedListener {
         void onResourceRemoved(Resource<?> removed);
@@ -22,24 +18,20 @@ public interface MemoryCache {
 
     /**
      * Returns the sum of the sizes of all the contents of the cache in bytes.
-     * <p>
-     * 返回已占用的缓存内存
      */
-    int getCurrentSize();
+    long getCurrentSize();
 
     /**
      * Returns the current maximum size in bytes of the cache.
-     * <p>
-     * 返回缓存分配到的最大内存
      */
-    int getMaxSize();
+    long getMaxSize();
 
     /**
      * Adjust the maximum size of the cache by multiplying the original size of the cache by the given
      * multiplier.
      * <p>
-     * If the size multiplier causes the size of the cache to be decreased, items will be evicted
-     * until the cache is smaller than the new size.
+     * <p> If the size multiplier causes the size of the cache to be decreased, items will be evicted
+     * until the cache is smaller than the new size. </p>
      *
      * @param multiplier A size multiplier >= 0.
      */
@@ -72,8 +64,6 @@ public interface MemoryCache {
 
     /**
      * Evict all items from the memory cache.
-     * <p>
-     * 从内存缓存中去除所有缓存项
      */
     void clearMemory();
